chore(watt): rename files
