+++ libs/ui-watt/src/lib/components/tabs/watt-tabs.component.scss
